+++ NEWS.md
@@ -49,6 +49,9 @@
 
 * Add ability to test `belongs_to` associations defined with `:inverse_of`.
 
+* Fix `rescue_from` matcher so that it does not raise an error when testing
+  a method handler which has been marked as protected or private.
+
 # v 2.5.0
 
 * Fix Rails/Test::Unit integration to ensure that the test case classes we are